refactor(network): separate network constructor from the rest
