+++ sn_networking/src/msg/mod.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::{error::Error, MsgResponder, NetworkEvent, SwarmDriver};
+use crate::{behaviour::SwarmDriver, error::Error, MsgResponder, NetworkEvent};
 use libp2p::request_response::{self, Message};
 use sn_protocol::messages::{Request, Response};
